move candidate assembly into a submodule
